"restructured application"
