--- src/App.css
@@ -1,92 +1,0 @@
-#cards {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  font-family: sans-serif;
-  font-weight: bold;
-  font-size: 1em;
-  text-align: center;
-}
-
-.singlecard {
-  display: flex;
-  flex-direction: row;
-  width: 40px;
-  height: 60px;
-  border: solid 2px black;
-  border-radius: 5px;
-  padding: 5px;
-  margin: 5px;
-  border-width: 1px;
-}
-
-.spades {
-  color: black;
-  background-image: url(.././public/spades.png);
-  background-repeat: no-repeat;
-  background-position: 50% 50%;
-  background-size: 30px;
-}
-
-.clubs {
-  color: black;
-  background-image: url(.././public/clubs.png);
-  background-repeat: no-repeat;
-  background-position: 50% 50%;
-  background-size: 30px;
-}
-
-.hearts {
-  color: red;
-  background-image: url(.././public/hearts.png);
-  background-repeat: no-repeat;
-  background-position: 50% 50%;
-  background-size: 30px;
-}
-
-.diamonds {
-  color: red;
-  background-image: url(.././public/diamonds.png);
-  background-repeat: no-repeat;
-  background-position: 50% 50%;
-  background-size: 30px;
-}
-
-.buttons {
-  padding: 10px;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-}
-
-button {
-  color: white;
-  text-shadow: 1px 1px 1px black;
-  width: 100px;
-  height: 40px;
-  border-radius: 20px;
-  border: 0;
-  font-family: sans-serif;
-  font-size: 1em;
-  font-weight: bold;
-  margin: 5px;
-  cursor: pointer;
-}
-
-button:hover {
-  text-shadow: none;
-  border: solid 2px black;
-  color: black;
-}
-
-.shuffle-deck-button {
-  background-color: #3aa137;
-}
-
-.full-deck-button {
-  background-color: #3489eb
-}
-
-.trick-deck-button {
-  background-color: orange;
-}
