+++ src/images/diamonds.png
