+++ src/styles/components/_cards.css
@@ -1,0 +1,43 @@
+.singlecard {
+  display: flex;
+  flex-direction: row;
+  width: 40px;
+  height: 60px;
+  border: solid 2px black;
+  border-radius: 5px;
+  padding: 5px;
+  margin: 5px;
+  border-width: 1px;
+}
+
+.spades {
+  color: black;
+  background-image: url(../../images/spades.png);
+  background-repeat: no-repeat;
+  background-position: 50% 50%;
+  background-size: 30px;
+}
+
+.clubs {
+  color: black;
+  background-image: url(../../images/clubs.png);
+  background-repeat: no-repeat;
+  background-position: 50% 50%;
+  background-size: 30px;
+}
+
+.hearts {
+  color: red;
+  background-image: url(../../images/hearts.png);
+  background-repeat: no-repeat;
+  background-position: 50% 50%;
+  background-size: 30px;
+}
+
+.diamonds {
+  color: red;
+  background-image: url(../../images/diamonds.png);
+  background-repeat: no-repeat;
+  background-position: 50% 50%;
+  background-size: 30px;
+}
