+++ src/styles/components/_deck.css
@@ -1,0 +1,10 @@
+#cards {
+  display: grid;
+  grid-template-columns: auto auto auto;
+  column-gap: 10px;
+  justify-content: center;
+  font-family: sans-serif;
+  font-weight: bold;
+  font-size: 1em;
+  text-align: center;
+}
