+++ src/styles/elements/_body.css
@@ -1,0 +1,6 @@
+body {
+  margin: 10px;
+  font-family: sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
